+++ pkgs/package_config/lib/src/package_config.dart
@@ -219,7 +219,7 @@ abstract class Package {
   /// The [packageUriRoot], if provided, must be either an absolute
   /// directory URI or a relative URI reference which is then resolved
   /// relative to [root]. It must then also be a subdirectory of [root],
-  /// or the same directory.
+  /// or the same directory, and must end with `/`.
   /// If [languageVersion] is supplied, it must be a valid Dart language
   /// version, which means two decimal integer literals separated by a `.`,
   /// where the integer literals have no leading zeros unless they are